修复部分主页因为依赖注入无法进入的问题 修复因为线程数量导致视频下载会卡住的问题 新增一小部分开发文档

+++ app/src/main/java/com/imcys/bilibilias/home/ui/model/UserNavDataModel.kt
@@ -212,7 +212,7 @@ data class UserNavDataModel(
             @SerializedName("coupon_due_time")
             val couponDueTime: Int, // 0
             @SerializedName("mid")
-            val mid: Int // 351201307
+            val mid: Long // 351201307
         ): Serializable
 
         data class WbiImg(
